requirements.txt
absl-py==0.9.0
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
astunparse==1.6.3
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
cmdline-csv2keychain==0.1.3
colorama==0.4.3
coloredlogs==14.0
configparser==5.0.0
crayons==0.3.1
cycler==0.10.0
distlib==0.3.0
Django==3.0.6
django-cors-headers==3.4.0
django-filter==2.2.0
djangorestframework==3.11.0
dlib==19.20.0
filelock==3.0.12
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
humanfriendly==8.2
idna==2.9
isort==4.3.21
Keras==2.4.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.2
oauthlib==3.1.0
opencv-contrib-python-headless==4.2.0.32
opencv-python==4.2.0.32
opt-einsum==3.2.1
pandas==1.0.3
Pillow==7.1.2
pipenv==2018.11.26
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pynput==1.6.8
pyobjc-core==6.2
pyobjc-framework-Cocoa==6.2
pyobjc-framework-Quartz==6.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
selenium==3.141.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
style==1.1.0
Tcl==0.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
tk==0.1.0
trunofficial==0.2.4
update==0.0.1
urllib3==1.25.9
virtualenv==20.0.18
virtualenv-clone==0.5.4
webdriver-manager==3.2.1
Werkzeug==1.0.1
wikipedia==1.4.0
wordcloud==1.7.0
wrapt==1.11.2
